+++ snail-boot/src/main/java/fun/easycode/snail/boot/core/CheckException.java
@@ -12,18 +12,12 @@ import lombok.EqualsAndHashCode;
 public class CheckException extends RuntimeException{
 
     /**
-     * 通用逻辑错误码
-     * @Author: xuzhen97
-     * @Date:  2023/03/22
-     */
-    public final static Integer DEFAULT_CODE = 40001;
-
-    /**
      * 错误码
-     * @Author: xuzhen97
-     * @Date:  2023/03/22
      */
     private final Integer code;
+    /**
+     * 错误信息
+     */
     private final String message;
 
     /**
@@ -32,7 +26,7 @@ public class CheckException extends RuntimeException{
      */
     public CheckException(String message){
         super(message);
-        this.code = DEFAULT_CODE;
+        this.code = R.DEFAULT_CODE;
         this.message = message;
     }
 
@@ -45,32 +39,5 @@ public class CheckException extends RuntimeException{
         super(message);
         this.code = code;
         this.message = message;
-    }
-
-    /**
-     * 返回错误信息
-     * @return 错误信息
-     */
-    public ErrorDto error(){
-        return new ErrorDto(code, message);
-    }
-
-    /**
-     * 返回错误信息
-     * @param code 错误码
-     * @param message 错误信息
-     * @return 错误信息
-     */
-    public static ErrorDto error(Integer code, String message){
-        return new ErrorDto(code, message);
-    }
-
-    /**
-     * 返回错误信息
-     * @param msg 错误信息
-     * @return 错误信息
-     */
-    public static ErrorDto error(String msg){
-        return new ErrorDto(DEFAULT_CODE, msg);
     }
 }