+++ changelog.md
@@ -1,6 +1,9 @@
+## 1.3.0-SNAPSHOT
+- refactor: CheckException 删除无用的方法，取消ErrorDto, 更换成R对象，统一正确和错误的接口返回值。
+
 ## 1.2.1-SNAPSHOT
 - fix: mapper xml 先是mapper后是模块名。
 
 ## 1.2.0-SNAPSHOT
 
-- 代码生成器修改，xml文件不放到dao下，独立到resources文件夹下的mapper, 目的减少copy-xmls配置。
+- 代码生成器修改，x你ml文件不放到dao下，独立到resources文件夹下的mapper, 目的减少copy-xmls配置。